# public/swagger-specs/services/transaction-scheduler/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "PegaSys Orchestrate Transaction API. Enables dynamic management of transactions.\nTransaction requests are an abstraction over schedules and jobs representing one or more transactions executed on the Blockchain network\nSchedules are ordered lists of jobs executed in a predefined sequence\nJobs represent a single transaction flow executed on the Blockchain network",
        "title": "Transaction Scheduler API",
        "contact": {
            "name": "Contact PegaSys Orchestrate",
            "url": "https://pegasys.tech/contact/",
            "email": "support@pegasys.tech"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "2.0"
    },
    "paths": {
...
        "/transactions": {
            "get": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    },
                    {
                        "JWTAuth": []
                    }
                ],
                "description": "Get a list of filtered transaction requests",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "summary": "Search transaction requests by provided filters",
                "parameters": [
                    {
                        "type": "array",
                        "items": {
                            "type": "string"
                        },
                        "collectionFormat": "csv",
                        "description": "List of idempotency keys",
                        "name": "idempotency_keys",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "List of transaction requests found",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/txscheduler.TransactionResponse"
                            }
                        }
                    },
                    "400": {
                        "description": "Invalid filter in the request",
                        "schema": {
                            "$ref": "#/definitions/httputil.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "$ref": "#/definitions/httputil.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/transactions/deploy-contract": {
            "post": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    },
                    {
                        "JWTAuth": []
                    }
                ],
                "description": "Creates and executes a new contract deployment request\nThe transaction can be private (Tessera, Orion).\nThe transaction can be a One Time Key transaction in 0 gas private networks",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "summary": "Creates and sends a new contract deployment",
                "parameters": [
                    {
                        "description": "Deployment transaction request",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/txscheduler.DeployContractRequest"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "params": {
                                            "allOf": [
                                                {
                                                    "$ref": "#/definitions/txscheduler.DeployContractParams"
                                                },
                                                {
                                                    "type": "object",
                                                    "properties": {
                                                        "gasPricePolicy": {
                                                            "allOf": [
                                                                {
                                                                    "$ref": "#/definitions/txscheduler.GasPriceParams"
                                                                },
                                                                {
                                                                    "type": "object",
                                                                    "properties": {
                                                                        "retryPolicy": {
                                                                            "$ref": "#/definitions/txscheduler.RetryParams"
                                                                        }
                                                                    }
                                                                }
                                                            ]
                                                        }
                                                    }
                                                }
                                            ]
                                        }
                                    }
                                }
                            ]
                        }
                    }
                ],
                "responses": {
                    "202": {
                        "description": "Created deployment transaction request",
                        "schema": {
                            "$ref": "#/definitions/txscheduler.TransactionResponse"
                        }
                    },
                    "400": {
                        "description": "Invalid request",
                        "schema": {
                            "$ref": "#/definitions/httputil.ErrorResponse"
                        }
                    },
                    "409": {
                        "description": "Already existing transaction",
                        "schema": {
                            "$ref": "#/definitions/httputil.ErrorResponse"
                        }
                    },
                    "422": {
                        "description": "Unprocessable parameters were sent",
                        "schema": {
                            "$ref": "#/definitions/httputil.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "$ref": "#/definitions/httputil.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/transactions/send": {
            "post": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    },
                    {
                        "JWTAuth": []
                    }
                ],
                "description": "Creates and executes a new smart contract transaction request\nThe transaction can be private (Tessera, Orion).\nThe transaction can be a One Time Key transaction in 0 gas private networks",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "summary": "Creates and sends a new contract transaction",
                "parameters": [
                    {
                        "description": "Contract transaction request",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/txscheduler.SendTransactionRequest"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "params": {
                                            "allOf": [
                                                {
                                                    "$ref": "#/definitions/txscheduler.TransactionParams"
                                                },
                                                {
                                                    "type": "object",
                                                    "properties": {
                                                        "gasPricePolicy": {
                                                            "allOf": [
                                                                {
                                                                    "$ref": "#/definitions/txscheduler.GasPriceParams"
                                                                },
                                                                {
                                                                    "type": "object",
                                                                    "properties": {
                                                                        "retryPolicy": {
                                                                            "$ref": "#/definitions/txscheduler.RetryParams"
                                                                        }
                                                                    }
                                                                }
                                                            ]
                                                        }
                                                    }
                                                }
                                            ]
                                        }
                                    }
                                }
                            ]
                        }
                    }
                ],
                "responses": {
                    "202": {
                        "description": "Created contract transaction request",
                        "schema": {
                            "$ref": "#/definitions/txscheduler.TransactionResponse"
                        }
                    },
                    "400": {
                        "description": "Invalid request",
                        "schema": {
                            "$ref": "#/definitions/httputil.ErrorResponse"
                        }
                    },
                    "409": {
                        "description": "Already existing transaction",
                        "schema": {
                            "$ref": "#/definitions/httputil.ErrorResponse"
                        }
                    },
                    "422": {
                        "description": "Unprocessable parameters were sent",
                        "schema": {
                            "$ref": "#/definitions/httputil.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "$ref": "#/definitions/httputil.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/transactions/send-raw": {
            "post": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    },
                    {
                        "JWTAuth": []
                    }
                ],
                "description": "Creates and executes a new raw transaction request",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "summary": "Creates and sends a raw transaction",
                "parameters": [
                    {
                        "description": "Raw transaction request",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/txscheduler.RawTransactionRequest"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "params": {
                                            "allOf": [
                                                {
                                                    "$ref": "#/definitions/txscheduler.RawTransactionParams"
                                                },
                                                {
                                                    "type": "object",
                                                    "properties": {
                                                        "retryPolicy": {
                                                            "$ref": "#/definitions/txscheduler.IntervalRetryParams"
                                                        }
                                                    }
                                                }
                                            ]
                                        }
                                    }
                                }
                            ]
                        }
                    }
                ],
                "responses": {
                    "202": {
                        "description": "Created raw transaction request",
                        "schema": {
                            "$ref": "#/definitions/txscheduler.TransactionResponse"
                        }
                    },
                    "400": {
                        "description": "Invalid request",
                        "schema": {
                            "$ref": "#/definitions/httputil.ErrorResponse"
                        }
                    },
                    "409": {
                        "description": "Already existing transaction",
                        "schema": {
                            "$ref": "#/definitions/httputil.ErrorResponse"
                        }
                    },
                    "422": {
                        "description": "Unprocessable parameters were sent",
                        "schema": {
                            "$ref": "#/definitions/httputil.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "$ref": "#/definitions/httputil.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/transactions/transfer": {
            "post": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    },
                    {
                        "JWTAuth": []
                    }
                ],
                "description": "Creates and executes a new transfer request",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "summary": "Creates and sends a transfer transaction",
                "parameters": [
                    {
                        "description": "Transfer transaction request",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/txscheduler.TransferRequest"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "params": {
                                            "allOf": [
                                                {
                                                    "$ref": "#/definitions/txscheduler.TransferParams"
                                                },
                                                {
                                                    "type": "object",
                                                    "properties": {
                                                        "gasPricePolicy": {
                                                            "allOf": [
                                                                {
                                                                    "$ref": "#/definitions/txscheduler.GasPriceParams"
                                                                },
                                                                {
                                                                    "type": "object",
                                                                    "properties": {
                                                                        "retryPolicy": {
                                                                            "$ref": "#/definitions/txscheduler.RetryParams"
                                                                        }
                                                                    }
                                                                }
                                                            ]
                                                        }
                                                    }
                                                }
                                            ]
                                        }
                                    }
                                }
                            ]
                        }
                    }
                ],
                "responses": {
                    "202": {
                        "description": "Created transfer transaction request",
                        "schema": {
                            "$ref": "#/definitions/txscheduler.TransactionResponse"
                        }
                    },
                    "400": {
                        "description": "Invalid request",
                        "schema": {
                            "$ref": "#/definitions/httputil.ErrorResponse"
                        }
                    },
                    "409": {
                        "description": "Already existing transaction",
                        "schema": {
                            "$ref": "#/definitions/httputil.ErrorResponse"
                        }
                    },
                    "422": {
                        "description": "Unprocessable parameters were sent",
                        "schema": {
                            "$ref": "#/definitions/httputil.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "$ref": "#/definitions/httputil.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/transactions/{uuid}": {
            "get": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    },
                    {
                        "JWTAuth": []
                    }
                ],
                "description": "Fetch a single transaction request by uuid",
                "produces": [
                    "application/json"
                ],
                "summary": "Fetch a transaction request by uuid",
                "parameters": [
                    {
                        "type": "string",
                        "description": "UUID of the transaction request",
                        "name": "uuid",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Transaction request found",
                        "schema": {
                            "$ref": "#/definitions/txscheduler.TransactionResponse"
                        }
                    },
                    "404": {
                        "description": "Transaction request not found",
                        "schema": {
                            "$ref": "#/definitions/httputil.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "$ref": "#/definitions/httputil.ErrorResponse"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
...
        "httputil.ErrorResponse": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "integer",
                    "example": 24000
                },
                "message": {
                    "type": "string",
                    "example": "invalid status update for the current job state"
                }
            }
        },
...
        "txscheduler.TransactionResponse": {
            "type": "object",
            "properties": {
                "chain": {
                    "type": "string",
                    "example": "myChain"
                },
                "createdAt": {
                    "type": "string",
                    "example": "2020-07-09T12:35:42.115395Z"
                },
                "idempotencyKey": {
                    "type": "string",
                    "example": "myIdempotencyKey"
                },
                "params": {
                    "type": "string"
                },
                "schedule": {
                    "type": "object",
                    "$ref": "#/definitions/txscheduler.ScheduleResponse"
                },
                "uuid": {
                    "type": "string",
                    "example": "b4374e6f-b28a-4bad-b4fe-bda36eaf849c"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "ApiKeyAuth": {
            "type": "apiKey",
            "name": "X-API-Key",
            "in": "header"
        },
        "JWTAuth": {
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}